refactor(FE): 불필요한 Container 컴포넌트 제거 (#146)
